// project/src/types/database.types.ts
export interface Database {
  public: {
    Tables: {
// ...
      resumes: {
        Row: {
          id: string
          created_at: string
          title: string
          version: string
          file_url: string
          is_active: boolean
          user_id: string
        }
        Insert: {
          id?: string
          created_at?: string
          title: string
          version: string
          file_url: string
          is_active?: boolean
          user_id: string
        }
        Update: {
          id?: string
          created_at?: string
          title?: string
          version?: string
          file_url?: string
          is_active?: boolean
          user_id?: string
        }
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
  }
}